Create index.html
This Widget is used on the eCommerce website to show products for sale. So here we made a responsive Shopping cart Widget in bootstrap 5 with the help of CSS and bootstrap components.

In this Widget, we used a simple grid and bootstrap card component with the helper class of bootstrap 5. Also, we modify bootstrap classes and use our own custom class on the responsive side. To make responsive we used CSS media quires with the required CSS properties we used. Also, we used media quires to make this Widget mobile-friendly.

Make it yours now by using it, downloading it, and please share it.